# pkg/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Ride-Hailing API
  description: |
    Comprehensive ride-hailing platform API supporting riders, drivers, and administrators.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the Authorization header.
    Obtain tokens via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting
    All endpoints are rate limited. Check `X-RateLimit-Limit`, `X-RateLimit-Remaining`,
    and `X-RateLimit-Reset` response headers.

    ## Idempotency
    For POST/PUT/PATCH requests, include an `Idempotency-Key` header to prevent duplicate processing.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Auth Service
  - url: http://localhost:8081
    description: Rides Service
  - url: http://localhost:8083
    description: Geo Service
  - url: http://localhost:8084
    description: Payments Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Ride:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rider_id:
          type: string
          format: uuid
        driver_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [requested, accepted, started, completed, cancelled]
        pickup_latitude:
          type: number
          format: double
        pickup_longitude:
          type: number
          format: double
        dropoff_latitude:
          type: number
          format: double
        dropoff_longitude:
          type: number
          format: double
        pickup_address:
          type: string
        dropoff_address:
          type: string
        fare:
          type: number
          format: double
        distance_km:
          type: number
          format: double
        duration_minutes:
          type: integer
        ride_type:
          type: string
        surge_multiplier:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
    LatitudeParam:
      name: latitude
      in: query
      required: true
      schema:
        type: number
        format: double
    LongitudeParam:
      name: longitude
      in: query
      required: true
      schema:
        type: number
        format: double
paths:
...
  /api/v1/rides:
    post:
      tags: [Rides]
      summary: Request a new ride
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickup_latitude, pickup_longitude, dropoff_latitude, dropoff_longitude]
              properties:
                pickup_latitude:
                  type: number
                  format: double
                pickup_longitude:
                  type: number
                  format: double
                dropoff_latitude:
                  type: number
                  format: double
                dropoff_longitude:
                  type: number
                  format: double
                pickup_address:
                  type: string
                dropoff_address:
                  type: string
                ride_type:
                  type: string
                promo_code:
                  type: string
      responses:
        '201':
          description: Ride requested
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Ride'
    get:
      tags: [Rides]
      summary: Get my rides
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: List of rides
  /api/v1/rides/{id}:
    get:
      tags: [Rides]
      summary: Get ride details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ride details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Ride'
  /api/v1/rides/{id}/cancel:
    post:
      tags: [Rides]
      summary: Cancel a ride
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Ride cancelled
  /api/v1/rides/{id}/rate:
    post:
      tags: [Rides]
      summary: Rate a completed ride
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '200':
          description: Rating submitted
  /api/v1/rides/surge-info:
    get:
      tags: [Rides]
      summary: Get surge pricing info for a location
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
      responses:
        '200':
          description: Surge pricing info

  # ==================== DRIVER ====================
  /api/v1/driver/rides/available:
    get:
      tags: [Driver]
      summary: Get available ride requests nearby
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Available rides
  /api/v1/driver/rides/{id}/accept:
    post:
      tags: [Driver]
      summary: Accept a ride request
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ride accepted
  /api/v1/driver/rides/{id}/start:
    post:
      tags: [Driver]
      summary: Start a ride (picked up passenger)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ride started
  /api/v1/driver/rides/{id}/complete:
    post:
      tags: [Driver]
      summary: Complete a ride
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ride completed

  # ==================== PAYMENTS ====================
...